pyproject.toml
[project]
name = "conways-pygame-of-life"
version = "0.1.0"
description = "Conway's Game of Life simulation made in Python using the Pygame-ce library"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce >= 2.5.6",
]
